feat: Save sequence on sort tab. (#375)
* feat: Save sequence on sort tab.

* fix: Edit with other client record.

+++ components/ADempiere/PanelDefinition/SortPanel.vue
@@ -24,7 +24,7 @@
     >
       <div class="board-column">
         <div class="board-column-header">
-          {{ this.$t('sequence.available') }} ({{ availableList.length }})
+          {{ this.$t('components.sequenceSort.available') }} ({{ availableList.length }})
         </div>
 
         <draggable
@@ -37,12 +37,14 @@
           <div
             v-for="(element, index) in availableList"
             :key="index"
-            class="board-item"
+            :disabled="isDifferentClientRecord"
+            :class="{ 'board-item': true, 'board-item-edit': element.isEditRow }"
           >
             <b>#{{ element[sortColumnName] }}</b>
             {{ element.DisplayColumn }}
 
             <i
+              :disabled="isDifferentClientRecord"
               class="el-icon-circle-plus-outline sort-add-item-icon"
               @click="addAtItem({ element, oldIndex: index })"
             />
@@ -57,7 +59,7 @@
     >
       <div class="board-column">
         <div class="board-column-header">
-          {{ this.$t('sequence.sequence') }} ({{ sequenceList.length }})
+          {{ this.$t('components.sequenceSort.sequence') }} ({{ sequenceList.length }})
         </div>
 
         <draggable
@@ -71,12 +73,14 @@
           <div
             v-for="(element, index) in sequenceList"
             :key="index"
-            class="board-item"
+            :disabled="isDifferentClientRecord"
+            :class="{ 'board-item': true, 'board-item-edit': element.isEditRow }"
           >
             <b>#{{ element[sortColumnName] }}</b>
             {{ element.DisplayColumn }}
 
             <i
+              :disabled="isDifferentClientRecord"
               class="el-icon-circle-close sort-remove-item-icon"
               @click="removeAtItem({ element, oldIndex: index })"
             />
@@ -95,13 +99,14 @@ import store from '@/store'
 // components and mixins
 import draggable from 'vuedraggable'
 
+// constants
+import { CLIENT } from '@/utils/ADempiere/constants/systemColumns'
+
 /**
  * Order or sequence panel based on the functionality of the `org.compiere.grid.VSortTab`
  * @see https://github.com/adempiere/adempiere/blob/develop/client/src/org/adempiere/controller/SortTabController.java
  * @see https://github.com/adempiere/adempiere/blob/develop/client/src/org/compiere/grid/VSortTab.java
  * TODO: Add change tracker to enable save button
- * TODO: Add change observer to show disable button
- * TODO: Highlight rows with changes
  * TODO: Add discartAtElementChange event to discard only current row
  * TODO: Implement search in local list and disable drag-and-drop
  */
@@ -170,6 +175,15 @@ export default defineComponent({
       }
     })
 
+    const oldRecordsList = computed(() => {
+      return store.getters.getTabSequenceOldRecordsList({
+        parentUuid: props.parentUuid,
+        containerUuid: props.containerUuid,
+        tabUuid: panelMetadata.value.uuid,
+        contextColumnNames: panelMetadata.value.contextColumnNames
+      })
+    })
+
     const availableList = computed({
       get() {
         return recordsList.value.filter(item => {
@@ -190,9 +204,22 @@ export default defineComponent({
       }
     })
 
+    const isDifferentClientRecord = computed(() => {
+      // client id value of record
+      const clientIdRecord = store.getters.getValueOfField({
+        parentUuid: props.parentUuid,
+        containerUuid: props.containerUuid,
+        columnName: CLIENT
+      })
+      // evaluate client id context with record
+      const sessionClientId = store.getters.getSessionContextClientId
+
+      return clientIdRecord !== sessionClientId
+    })
+
     const dragOptions = computed(() => {
       return {
-        disabled: false
+        disabled: isDifferentClientRecord.value
       }
     })
 
@@ -200,6 +227,33 @@ export default defineComponent({
       return arrayToSort.sort((itemA, itemB) => {
         return itemA[orderBy] - itemB[orderBy]
       })
+    }
+
+    function getOldRow(uuid) {
+      const oldRow = oldRecordsList.value.find(oldSequenceItem => {
+        return oldSequenceItem.UUID === uuid
+      })
+      return {
+        ...oldRow
+      }
+    }
+
+    function isChanged({ oldRow, newRow }) {
+      const oldYesNo = oldRow[includedColumnName.value]
+      const newYesNo = newRow[includedColumnName.value]
+      // included column is changed
+      if (oldYesNo !== newYesNo) {
+        return true
+      }
+
+      const oldSequence = oldRow[sortColumnName.value]
+      const newSequence = newRow[sortColumnName.value]
+      // sequence is changed
+      if (oldSequence !== newSequence) {
+        return true
+      }
+
+      return false
     }
 
     /**
@@ -219,10 +273,20 @@ export default defineComponent({
 
       const dataSequence = recordsList.value.map(itemSequence => {
         if (itemSequence.UUID === element.UUID) {
+          const oldRow = getOldRow(itemSequence.UUID)
+          element.isEditRow = isChanged({
+            oldRow,
+            newRow: element
+          })
           return element
         }
         if (newSequence <= itemSequence[orderColumn]) {
           itemSequence[orderColumn] = itemSequence[orderColumn] + 10
+          const oldRow = getOldRow(itemSequence.UUID)
+          itemSequence.isEditRow = isChanged({
+            newRow: itemSequence,
+            oldRow
+          })
         }
         return itemSequence
       })
@@ -254,19 +318,39 @@ export default defineComponent({
           // moved to down
           if (itemSequence.UUID === element.UUID) {
             itemSequence[orderColumn] = (newIndex + 1) * 10
+            const oldRow = getOldRow(itemSequence.UUID)
+            itemSequence.isEditRow = isChanged({
+              newRow: itemSequence,
+              oldRow
+            })
             return itemSequence
           }
           if (indexEnabledSequence >= oldIndex && indexEnabledSequence < newIndex) {
             itemSequence[orderColumn] = (indexEnabledSequence + 1) * 10
+            const oldRow = getOldRow(itemSequence.UUID)
+            itemSequence.isEditRow = isChanged({
+              newRow: itemSequence,
+              oldRow
+            })
           }
         } else {
           // moved to up
           if (itemSequence.UUID === element.UUID) {
             itemSequence[orderColumn] = (newIndex + 1) * 10
+            const oldRow = getOldRow(itemSequence.UUID)
+            itemSequence.isEditRow = isChanged({
+              newRow: itemSequence,
+              oldRow
+            })
             return itemSequence
           }
           if (indexEnabledSequence < oldIndex && indexEnabledSequence >= newIndex) {
             itemSequence[orderColumn] += 10
+            const oldRow = getOldRow(itemSequence.UUID)
+            itemSequence.isEditRow = isChanged({
+              newRow: itemSequence,
+              oldRow
+            })
           }
         }
         indexEnabledSequence++
@@ -295,11 +379,24 @@ export default defineComponent({
         if (itemSequence.UUID === element.UUID) {
           itemSequence[includedColumn] = false // !itemSequence[includedColumn]
           itemSequence[orderColumn] = 0
+          const oldRow = getOldRow(itemSequence.UUID)
+          itemSequence.isEditRow = isChanged({
+            newRow: itemSequence,
+            oldRow
+          })
           return itemSequence
         }
+
         if (itemSequence[orderColumn] > oldSequence && itemSequence[orderColumn] > 0) {
           itemSequence[orderColumn] = itemSequence[orderColumn] - 10
+
+          const oldRow = getOldRow(itemSequence.UUID)
+          itemSequence.isEditRow = isChanged({
+            newRow: itemSequence,
+            oldRow
+          })
         }
+
         return itemSequence
       })
 
@@ -316,6 +413,10 @@ export default defineComponent({
      * @param {number} oldIndex: the index of the element before remove
      */
     function addAtItem({ element, oldIndex }) {
+      if (isDifferentClientRecord.value) {
+        // is another customer does not change
+        return
+      }
       const newIndex = sequenceList.value.length
       addItem({
         element,
@@ -329,6 +430,10 @@ export default defineComponent({
      * @param {number} oldIndex: the index of the element before remove
      */
     function removeAtItem({ element, oldIndex }) {
+      if (isDifferentClientRecord.value) {
+        // is another customer does not change
+        return
+      }
       deleteItem({
         element,
         oldIndex
@@ -360,6 +465,7 @@ export default defineComponent({
       dragOptions,
       panelMetadata,
       includedColumnName,
+      isDifferentClientRecord,
       sortColumnName,
       sequenceList,
       availableList,
@@ -401,6 +507,8 @@ export default defineComponent({
       justify-content: flex-start;
       flex-direction: column;
       align-items: center;
+      padding-left: 5px;
+      padding-right: 5px;
 
       .board-item {
         cursor: pointer;
@@ -413,6 +521,14 @@ export default defineComponent({
         padding: 0px 10px;
         box-sizing: border-box;
         box-shadow: 0px 1px 3px 0 rgba(0, 0, 0, 0.2);
+
+        font-size: 1.2em;
+
+        &.board-item-edit {
+          -webkit-box-shadow: 0px 0px 6px 0px rgba(21,255,0,1);
+          -moz-box-shadow: 0px 0px 6px 0px rgba(21,255,0,1);
+          box-shadow: 0px 0px 6px 0px rgba(21,255,0,1);
+        }
       }
     }
   }
@@ -460,6 +576,12 @@ export default defineComponent({
           font-weight: bold;
         }
       }
+    }
+
+    .board-item[disabled=disabled],
+    .sort-remove-item-icon[disabled=disabled],
+    .sort-add-item-icon[disabled=disabled] {
+      cursor: not-allowed;
     }
   }
 }
